+++ app/src/main/java/com/artem/animationjikan/ui/tab/home/home.kt
@@ -1,11 +1,9 @@
-package com.artem.animationjikan.ui.tab
+package com.artem.animationjikan.ui.tab.home
 
-import androidx.compose.foundation.Indication
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxHeight
@@ -29,6 +27,7 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import coil3.compose.AsyncImage
+import com.artem.animationjikan.ui.tab.home.components.RecommendPager
 
 @Composable
 fun HomeTab(modifier: Modifier = Modifier) {
@@ -38,7 +37,6 @@ fun HomeTab(modifier: Modifier = Modifier) {
 
     /// TODO  PageCount는 API 연동시 해당 Length에 맞게 변경하기
     val pageCount = 5
-    val pagerState = rememberPagerState(pageCount = { pageCount })
 
     Box(
         modifier
@@ -63,60 +61,11 @@ fun HomeTab(modifier: Modifier = Modifier) {
                 }
             }
 
-            HorizontalPager(
-                state = pagerState
-            ) { page ->
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .fillMaxHeight(0.6f)
-                        .background(color = getPageColor(page)),
-                    contentAlignment = Alignment.Center
-                ) {
-                    AsyncImage(
-                        model = "https://cdn.myanimelist.net/images/anime/1015/138006.jpg",
-                        contentDescription = "Poster",
-                        modifier = Modifier
-                            .fillMaxHeight()
-                            .aspectRatio(0.75f),
-                        contentScale = ContentScale.Crop
-                    )
-                }
-            }
-
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(
-                        vertical = 16.dp
-                    ),
-                horizontalArrangement = Arrangement.Center
-            ) {
-                repeat(pageCount) { iteration ->
-                    val color = if (pagerState.currentPage == iteration)
-                        Color(0xFFE50913) else Color.LightGray
-
-                    Box(
-                        modifier = Modifier
-                            .padding(horizontal = 4.dp)
-                            .size(8.dp)
-                            .background(color, CircleShape)
-                    )
-                }
-            }
+            RecommendPager(pageCount = pageCount)
         }
     }
 }
 
-private fun getPageColor(page: Int): Color {
-    return when (page % 5) {
-        0 -> Color(0xFFEF5350)
-        1 -> Color(0xFF66BB6A)
-        2 -> Color(0xFF26C6DA)
-        3 -> Color(0xFF7E57C2)
-        else -> Color(0xFFFFCA28)
-    }
-}
 
 @Composable
 @Preview
