+++ app/src/main/java/com/artem/animationjikan/ui/tab/like/like.kt
@@ -1,4 +1,4 @@
-package com.artem.animationjikan.ui.tab
+package com.artem.animationjikan.ui.tab.like
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
