+++ app/src/main/java/com/artem/animationjikan/util/router/NavRoutes.kt
@@ -2,6 +2,8 @@ package com.artem.animationjikan.util.router
 
 sealed class NavRoutes(val router: String) {
     data object Home : NavRoutes("home")
-    data object Like : NavRoutes("like")
-    data object Search : NavRoutes("search")
+    data object AnimationDetail : NavRoutes("animationDetail")
+    data object MangaDetail : NavRoutes("MangaDetail")
+    data object CharacterDetail : NavRoutes("CharacterDetail")
+    data object VoiceActorDetail : NavRoutes("VoiceActorDetail")
 }
